+++ projects/core/src/lib/refresher/refresher-content/refresher-content.component.ts
@@ -8,74 +8,19 @@ import {
   EventEmitter,
   Renderer2
 } from '@angular/core';
-import {
-  animate,
-  style,
-  transition,
-  trigger,
-  state,
-  AnimationBuilder
-} from '@angular/animations';
-import { FivLoadingSpinner } from '../../loading-spinner/loading-spinner.component';
+import { animate, style, AnimationBuilder } from '@angular/animations';
+import { FivSpinner } from '../../spinner/spinner.component';
 
 @Component({
   selector: 'fiv-refresher-content',
   templateUrl: './refresher-content.component.html',
-  styleUrls: ['./refresher-content.component.scss'],
-  animations: [
-    trigger('fabAnim', [
-      transition('void => *', [
-        style({ transform: 'scale(0)' }),
-        animate('250ms ease-out')
-      ]),
-      transition('* => void', [
-        animate('250ms ease-in', style({ transform: 'scale(0)' }))
-      ])
-    ]),
-    trigger('spinnerAnim', [
-      transition('void => *', [
-        style({ opacity: '0' }),
-        animate('250ms ease-out')
-      ]),
-      transition('* => void', [
-        animate('250ms ease-in', style({ opacity: '0' }))
-      ])
-    ]),
-    trigger('rotateAnim', [
-      transition('normal => rotate', [animate('125ms ease-out')]),
-      transition('rotate => normal', [animate('125ms ease-in')]),
-      state('rotate', style({ opacity: '0', transform: 'rotateZ(45deg)' })),
-      state('normal', style({ opacity: '1', transform: 'rotateZ(0deg)' }))
-    ]),
-    trigger('fillAnim', [
-      transition('* => fill', [
-        style({
-          'stroke-dasharray': 180,
-          'stroke-dashoffset': 90,
-          transformOrigin: 'center',
-          stroke: '#DE3E35'
-        }),
-        animate('1400ms ease-out')
-      ]),
-      state(
-        'fill',
-        style({
-          'stroke-dasharray': 315,
-          'stroke-dashoffset': 0,
-          transformOrigin: 'center',
-          stroke: '#1B9A59',
-          opacity: 0
-        })
-      )
-    ])
-  ]
+  styleUrls: ['./refresher-content.component.scss']
 })
 export class FivRefresherContent implements OnInit {
   @Input() icon: string;
   @Input() spinColor: string;
   @Input() fabColor: string;
   @Input() iconColor = '#000';
-  @Input() checkmark = false;
   @Input() disabled = false;
   _progress = 0;
   @Input() set progress(progress) {
@@ -107,7 +52,7 @@ export class FivRefresherContent implements OnInit {
   >();
   @Output() fivProgress: EventEmitter<number> = new EventEmitter<number>();
 
-  @ViewChild('spinner') spinner: FivLoadingSpinner;
+  @ViewChild('spinner') spinner: FivSpinner;
   @ViewChild('spinner', { read: ElementRef }) spinnerRef: ElementRef;
   @ViewChild('background') background: ElementRef;
 
@@ -129,16 +74,7 @@ export class FivRefresherContent implements OnInit {
   }
 
   unload() {
-    this.spinner.stopSpinning();
-  }
-
-  complete() {
-    // this.spinner.completeIn(500);
-    if (this.checkmark) {
-      this.iconState = 'rotate';
-    } else {
-      this.postComplete();
-    }
+    this.spinner.stop();
   }
 
   show() {
@@ -163,7 +99,7 @@ export class FivRefresherContent implements OnInit {
       const transform = `scale(1)`;
       const animation = this.builder.build([
         style({ transform: transform }),
-        animate('175ms ease-in', style({ transform: 'scale(0)' }))
+        animate('175ms ease-in', style({ transform: `scale(0)` }))
       ]);
 
       const player = animation.create(this.background.nativeElement);
@@ -176,26 +112,7 @@ export class FivRefresherContent implements OnInit {
     }
   }
 
-  fillAnimationDone() {
-    if (this.checkmark) {
-      this.iconState = 'rotate';
-    } else {
-      this.postComplete();
-    }
-  }
-
-  changeIconAndReveal(event, icon: string) {
-    if (event.fromState === 'normal') {
-      this.icon = icon;
-      this.iconState = 'normal';
-    } else {
-      if (event.fromState === 'rotate') {
-        this.postComplete();
-      }
-    }
-  }
-
-  postComplete() {
+  complete() {
     this.unload();
     this.fivComplete.emit(this);
   }