+++ projects/core/src/public_api.ts
@@ -25,7 +25,7 @@ export * from './lib/refresher/refresher.module';
 export * from './lib/fab/fab.module';
 export * from './lib/loading-progress-bar/loading-progress-bar.module';
 export * from './lib/loading-progress-bar/loading-progress-bar.module';
-export * from './lib/loading-spinner/loading-spinner.module';
+export * from './lib/spinner/spinner.module';
 export * from './lib/network-status/network-status.module';
 export * from './lib/overlay/overlay.module';
 export * from './lib/password-input/password-input.module';